refactor(telegram): extract config mention observation
